requirements.txt
absl-py==2.1.0
accelerate==0.20.3
aiohttp==3.9.5
aiosignal==1.3.1
alembic==1.14.0
annotated-types==0.7.0
antlr4-python3-runtime==4.9.3
appdirs==1.4.4
asciitree==0.3.3
asttokens==2.4.1
async-timeout==4.0.3
attrs==24.2.0
autocommand==2.2.2
autopage==0.5.2
backports.tarfile==1.2.0
bioframe==0.7.2
biopython==1.83
biotite==0.37.0
biotraj==1.2.2
certifi==2024.8.30
cfgv==3.4.0
charset-normalizer==3.4.0
click==8.1.7
cliff==4.7.0
cmaes==0.11.1
cmd2==2.5.6
colorlog==6.9.0
contourpy==1.3.1
cooler==0.9.3
cooltools==0.7.1
cycler==0.12.1
Cython==3.0.11
cytoolz==1.0.0
datasets==2.1.0
decorator==5.1.1
dgl==2.4.0+cu121
diffusers==0.31.0
dill==0.3.8
dirichlet==0.9
distlib==0.3.9
docker-pycreds==0.4.0
easydict==1.13
einops==0.7.0
exceptiongroup==1.2.2
executing==2.1.0
fair-esm==2.0.0
fairscale==0.4.13
filelock==3.16.1
fisher-flow==1.0.0
fonttools==4.55.0
frozenlist==1.4.1
fsspec==2024.10.0
geoopt==0.5.0
gitdb==4.0.11
GitPython==3.1.43
greenlet==3.1.1
grpcio==1.68.0
h5py==3.12.1
huggingface-hub==0.26.2
hydra-colorlog==1.2.0
hydra-core==1.3.2
hydra-optuna-sweeper==1.2.0
identify==2.6.2
idna==3.10
imageio==2.36.0
importlib_metadata==8.5.0
importlib_resources==6.4.0
inflect==7.3.1
iniconfig==2.0.0
ipdb==0.13.13
ipython==8.29.0
jaraco.collections==5.1.0
jaraco.context==5.3.0
jaraco.functools==4.0.1
jaraco.text==3.12.1
jedi==0.19.2
Jinja2==3.1.4
joblib==1.4.2
kiwisolver==1.4.7
lazy_loader==0.4
lightning==2.4.0
lightning-utilities==0.11.8
llvmlite==0.43.0
lmdb==1.5.1
loralib==0.1.2
Mako==1.3.6
Markdown==3.7
markdown-it-py==3.0.0
MarkupSafe==3.0.2
matplotlib==3.9.2
matplotlib-inline==0.1.7
mdurl==0.1.2
more-itertools==10.3.0
mpmath==1.3.0
msgpack==1.1.0
multidict==6.0.5
multiprocess==0.70.16
natsort==8.4.0
ncls==0.0.68
networkx==3.4.2
nodeenv==1.9.1
numba==0.60.0
numpy==1.26.4
nvidia-cublas-cu12==12.1.3.1
nvidia-cuda-cupti-cu12==12.1.105
nvidia-cuda-nvrtc-cu12==12.1.105
nvidia-cuda-runtime-cu12==12.1.105
nvidia-cudnn-cu12==8.9.2.26
nvidia-cufft-cu12==11.0.2.54
nvidia-curand-cu12==10.3.2.106
nvidia-cusolver-cu12==11.4.5.107
nvidia-cusparse-cu12==12.1.0.106
nvidia-nccl-cu12==2.18.1
nvidia-nvjitlink-cu12==12.6.77
nvidia-nvtx-cu12==12.1.105
omegaconf==2.3.0
optuna==2.10.1
packaging==24.2
pandas==2.2.3
parso==0.8.4
patsy==1.0.1
pbr==6.1.0
pexpect==4.9.0
pillow==11.0.0
platformdirs==4.3.6
plotly==5.24.1
pluggy==1.5.0
POT==0.9.3
pre_commit==4.0.1
prettytable==3.12.0
prompt_toolkit==3.0.48
protobuf==4.25.5
psutil==6.1.0
ptyprocess==0.7.0
pure_eval==0.2.3
pyarrow==16.1.0
pyBigWig==0.3.22
pydantic==2.5.3
pydantic_core==2.14.6
pyDeprecate==0.3.2
pyfaidx==0.8.1.3
pyg-lib==0.4.0+pt21cu121
Pygments==2.18.0
pyparsing==3.2.0
pyperclip==1.9.0
pyranges==0.0.129
pyrootutils==1.0.4
pytabix==0.1
pytest==8.3.3
python-dateutil==2.9.0.post0
python-dotenv==1.0.1
pytorch-lightning==2.4.0
pytz==2024.2
PyYAML==6.0.2
rdkit==2023.9.5
regex==2024.11.6
requests==2.32.3
responses==0.18.0
rich==13.9.4
rootutils==1.0.7
sacremoses==0.0.53
safetensors==0.4.5
schedulefree==1.2.5
scikit-image==0.24.0
scikit-learn==1.5.2
scipy==1.13.1
seaborn==0.12.2
selene-sdk==0.4.4
sentry-sdk==2.18.0
setproctitle==1.3.4
simplejson==3.19.3
six==1.16.0
smmap==5.0.1
sorted-nearest==0.0.39
SQLAlchemy==2.0.36
stack-data==0.6.3
statsmodels==0.14.4
stevedore==5.3.0
sympy==1.13.3
tabulate==0.9.0
tenacity==9.0.0
tensorboard==2.18.0
tensorboard-data-server==0.7.2
threadpoolctl==3.5.0
tifffile==2024.9.20
tmtools==0.2.0
tokenizers==0.19.1
tomli==2.1.0
toolz==1.0.0
torch==2.1.0
torch-cluster==1.6.3+pt21cu121
torch-ema==0.3
torch-scatter==2.1.2+pt21cu121
torch-sparse==0.6.18+pt21cu121
torch-spline-conv==1.2.2+pt21cu121
torch_geometric==2.5.3
torchdata==0.7.0
torchdiffeq==0.2.3
torchmetrics==1.6.0
torchtext==0.16.0
torchvision==0.16.0
tqdm==4.67.0
traitlets==5.14.3
transformers==4.40.2
triton==2.1.0
typeguard==4.3.0
typing_extensions==4.12.2
tzdata==2024.2
urllib3==2.2.3
virtualenv==20.27.1
wandb==0.18.7
wcwidth==0.2.13
Werkzeug==3.1.3
xxhash==3.4.1
yarl==1.9.4
zipp==3.21.0
